music-agent/requirements.txt
textual>=0.80.0
SQLAlchemy>=2.0
PyYAML>=6.0
requests>=2.31.0
pycryptodome>=3.20.0
